% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{alignThreshold}
\alias{alignThreshold}
\title{Aligns a threshold from a reference to a test density.}
\usage{
alignThreshold(densX, densY, refDensX, refDensY, refThreshold)
}
\arguments{
\item{densX}{Density x values of the test density.}

\item{densY}{Density y values of the test density.}

\item{refDensX}{Density x values of the reference density.}

\item{refDensY}{Density y values of the reference density.}

\item{refThreshold}{Threshold of the reference density.}
}
\value{
The predicted threshold in the test density.
}
\description{
Aligns a threshold from a reference to a test density.
}

