% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{derivativeDtwDistanceMatrix}
\alias{derivativeDtwDistanceMatrix}
\title{Computes a distance matrix for derivative DTW, given two densities.}
\usage{
derivativeDtwDistanceMatrix(x, y)
}
\arguments{
\item{x}{First density of length n.}

\item{y}{Second density of length n.}
}
\value{
A n-by-n matrix with pairwise derivative distances.
}
\description{
Computes a distance matrix for derivative DTW, given two densities.
}

