% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{dtwDistanceMatrices}
\alias{dtwDistanceMatrices}
\title{Computes DTW distance matrices for channel A and B, respectively.}
\usage{
dtwDistanceMatrices(tr, cl)
}
\arguments{
\item{tr}{A LearningSet object.}

\item{cl}{A compute cluster as obtained by parallel::makeCluster()}
}
\value{
A list with keys dA and dB indicating the distance matrices for channel A and B.
}
\description{
Each matrix is of size nxn where n is the number of samples.
This implementation is parallelized. Only the upper triangle matrix is computed and then mirrored.
}

