% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{dtwDistanceMatrix}
\alias{dtwDistanceMatrix}
\title{Computes a euclidean distance matrix for two given densities.}
\usage{
dtwDistanceMatrix(x, y)
}
\arguments{
\item{x}{First density of length n.}

\item{y}{Second density of length m.}
}
\value{
A n-by-m matrix with pairwise euclidean distances.
}
\description{
Computes a euclidean distance matrix for two given densities.
}

