% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{dtwMain}
\alias{dtwMain}
\title{Global proxy function which modularizes the parameters for DTW}
\usage{
dtwMain(a, b, ...)
}
\arguments{
\item{a}{First density of length n.}

\item{b}{Second density of length n.}
}
\value{
A dtw object as returned by dtw::dtw.
}
\description{
This function exists to ensure consistent use throughout flowLearn. 
To Change DTW parameters, change this function.
Calls dtw::dtw with parameters suitable for flowLean.
}

