% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{estimateDensity}
\alias{estimateDensity}
\title{Estimates a smoothed density from a given vector of numbers.}
\usage{
estimateDensity(data, n)
}
\arguments{
\item{data}{The numbers to calculated the density for.}

\item{n}{Number of features for the density.}
}
\value{
The estimated density.
}
\description{
Estimates a smoothed density from a given vector of numbers.
}

