% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flAdd}
\alias{flAdd}
\title{Adds one row (i.e. one density) to an existing DensityData object.}
\usage{
flAdd(obj, fcs, population, channelIdx, parentDensX, parentDensY,
  gate.low = NaN, gate.high = NaN)
}
\arguments{
\item{obj}{The existing DensityData object.}

\item{fcs}{The fcs file name of the new density.}

\item{population}{The population of the new density.}

\item{channelIdx}{The channel index of the new density.}

\item{parentDensX}{The parent population's density x-values of the gated population.}

\item{parentDensY}{The parent population's density y-values of the gated population.}

\item{gate.low}{The lower threshold of the gated channel (optional, default: NaN).}

\item{gate.high}{The upper threshold of the gated channel (optional, default: NaN).}
}
\value{
The supplied DensityData obj, including a new row with the supplied parameters.
}
\description{
Adds one row (i.e. one density) to an existing DensityData object.
}
\examples{
densdat <- flInit(new('DensityData')) # create empty DensityData object
densdat <- flAdd(densdat, 'mine.fcs', 'CD43+', 1, 1:512, runif(n = 512, min = 1, max = 10), 2, 7) # add one channel density

}
