% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flAddFlowFrame}
\alias{flAddFlowFrame}
\title{Adds densities from a flowFrame that represents a parent cell population.}
\usage{
flAddFlowFrame(obj, parentFlowFrame, channelIndices, populationName,
  fcsName = NULL)
}
\arguments{
\item{obj}{A DensityData object.}

\item{parentFlowFrame}{An object of type flowFrame representing the parent population.}

\item{channelIndices}{A vector of indices of channels to use for gating the child population.}

\item{populationName}{The name of the child population.}

\item{fcsName}{The different name for the FCS file. By default parentFlowFrame@description$"$FIL" is used.}
}
\value{
The new DensityData object containing one density for each channel specified.
}
\description{
Adds densities from a flowFrame that represents a parent cell population.
}
\examples{
densdat <- flInit(new('DensityData'))
flSampleFlowFrame <- readRDS(gzcon(url('https://raw.githubusercontent.com/mlux86/flowLearn/master/extra/data/flSampleFlowFrame.rds')))
densdat <- flAddFlowFrame(densdat, flSampleFlowFrame, c(1,2), 'Granulocyte Pre')

}
