% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flAt}
\alias{flAt}
\title{Returns a single density from the data slot at the specified row index.}
\usage{
flAt(obj, idx)
}
\arguments{
\item{obj}{The DensityData object to be filtered.}

\item{idx}{The row index of the internal "data" data frame.}
}
\value{
A new DensityData object with only one row, given by the specified index.
}
\description{
Returns a single density from the data slot at the specified row index.
}
\examples{
dd <- flAt(flSampleDensdat, 42) # grab 42nd entry of flSampleDensdat
      
}
