% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flConcat}
\alias{flConcat}
\title{Concatenates two DensityData objects.}
\usage{
flConcat(obj1, obj2)
}
\arguments{
\item{obj1}{The first DensityData object.}

\item{obj2}{The second DensityData object.}
}
\value{
A new DensityData object consisting of the internal "data" data frames of obj1 and obj2.
}
\description{
Concatenates two DensityData objects.
}
\examples{
print(flSize(flSampleDensdat))
dd <- flConcat(flSampleDensdat, flSampleDensdat)
print(flSize(dd))

}
