% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flData}
\alias{flData}
\title{Getter for the "data" slot}
\usage{
flData(obj)
}
\arguments{
\item{obj}{A DensityData object.}
}
\value{
The internal data frame "data"
}
\description{
Getter for the "data" slot
}
\examples{
flData(flSampleDensdat)
      
}
