% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flDerivativeDtwDistanceMatrix}
\alias{flDerivativeDtwDistanceMatrix}
\title{Given two R density objects, calculates a distance matrix based on the derivative DTW distance.}
\usage{
flDerivativeDtwDistanceMatrix(densA, densB)
}
\arguments{
\item{densA}{The first density object, such as from R's density function.}

\item{densB}{The second density object, such as from R's density function.}
}
\value{
The Derivative Dynamic Time Warping distance matrix between densA and densB.
}
\description{
Given two R density objects, this method uses the Derivative Dynamic Time Warping distance to calculate a distance matrix which can be used for alignment with DTW.
}
\examples{
densA <- flGetDensity(flAt(flSampleDensdat, 42))
densB <- flGetDensity(flAt(flSampleDensdat, 43))
str(flDerivativeDtwDistanceMatrix(densA, densB))

}
