% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flDtwMain}
\alias{flDtwMain}
\title{Wrapper function to carry out DDTW for flowLearn.}
\usage{
flDtwMain(densA, densB, ...)
}
\arguments{
\item{densA}{The first density object, such as from R's density function.}

\item{densB}{The second density object, such as from R's density function.}

\item{...}{Optional parameters given to dtw::dtw}
}
\value{
The Derivative Dynamic Time Warping object from the dtw package.
}
\description{
Wrapper function to carry out Derivative Dynamic Time Warping for flowLearn. It uses Derivative DTW with a DTW specific step pattern, which is "typeIds" from the dtw package.
}
\examples{
dd <- flFind(flSampleDensdat, 'population == "cd3tcell" & channelIdx == 1')
dtwObj <- flDtwMain(flGetDensity(flAt(dd, 1)), flGetDensity(flAt(dd, 2)))
plot(dtwObj)

}
