% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flEstimateDensity}
\alias{flEstimateDensity}
\title{Estimates a density on a given data vector.}
\usage{
flEstimateDensity(data, n)
}
\arguments{
\item{data}{The data vector for which the density is calculated.}

\item{n}{The number of density features.}
}
\value{
The calculated density object.
}
\description{
This method calculates densities used by flowLearn. It uses the default stats::density function and smoothens the result using stats::spline.smooth.
}
\examples{
x <- seq(-pi, pi, by = 0.01)
y <- sin(x)
noise <- runif(n = length(y), min = 0, max = 1)
noisy_y <- y + noise
par(mfrow = c(2,1))
plot(noisy_y)
plot(flEstimateDensity(noisy_y, 512))

}
