% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{flEvalF1ScoreFCS}
\alias{flEvalF1ScoreFCS}
\title{Given a DensityData object with existing thresholds, calculates precision, recall, F1-score, true and predicted cell proportions for one FCS file.}
\usage{
flEvalF1ScoreFCS(densdat, fcs, population, trueAssignments, parentExprs,
  negate = FALSE)
}
\arguments{
\item{densdat}{The full DensityData object with density and densities and set/predicted thresholds.}

\item{fcs}{The name of the FCS file to consider. This is needed to filter the relevant entries in densdat.}

\item{population}{The name of the population to gate. This is needed to filter the relevant entries in densdat.}

\item{trueAssignments}{A vector of length <number of cells>, indicating true cell memberships of the target population.}

\item{parentExprs}{A n*d expression matrix of the parent population where n is the number of cells and d is the number of channels. The channel indices correspond to the channelIdx values in densdat.}

\item{negate}{Whether the gated population should be negated/inverted, i.e. not the cells within the gate but out of it are considered.}
}
\value{
A data.frame with precision, recall, f1, trueProportion, predictedProportion.
}
\description{
This method gates each cell and compares the result to a true vector of true cell assignments. It returns its performance in the form of various parameters.
}
\examples{
# In this example, we evaluate true against the true gate assignments, hence it should give perfect performance, i.e. F_1 = 1

f <- unique(flData(flSampleDensdat)$fcs)[[1]]
flEvalF1ScoreFCS(flSampleDensdat, f, 'bcell', flSampleBcellEvaluationData[[f]]$gateAssignments, flSampleBcellEvaluationData[[f]]$parentExprs, FALSE)


}
