% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flFind}
\alias{flFind}
\title{Returns a new DensityData object with a subset of entries of the supplied obj.}
\usage{
flFind(obj, fcs, population, channelIdx)
}
\arguments{
\item{fcs}{The FCS file to filter for.}

\item{population}{The population to filter for.}

\item{channelIdx}{The channel to filter for.}
}
\value{
A new DensityData object with a subset of entries, defined by mysubset.
}
\description{
Returns a new DensityData object with a subset of entries of the supplied obj.
}
\examples{
# Get all densities for Population "hfa" and the first channel
dd <- flFind(flSampleDensdat, population = 'hfa', channelIdx = 1)

}
