% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flGetGate}
\alias{flGetGate}
\title{Returns a matrix with two columns which represent lower and upper thresholds of the given DensityData object.}
\usage{
flGetGate(obj)
}
\arguments{
\item{obj}{The DensityData object.}
}
\value{
A matrix with columns "gate.low" and "gate.high", representing lower and upper thresholds on all rows/channels present obj.
}
\description{
Returns a matrix with two columns which represent lower and upper thresholds of the given DensityData object.
}
\examples{
dd <- flAt(flSampleDensdat, 42) # grab density at row index 42 
dens <- flGetDensity(dd) # 
gt <- flGetGate(dd) # returns vector with "gate.low" and "gate.high"
flPlotDensThresh(dens, gt)
      
}
