% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{flGetGateAssignments}
\alias{flGetGateAssignments}
\title{Given a DensityData object with existing thresholds, returns a logical vector indicating cell memberships.}
\usage{
flGetGateAssignments(densdat, exprs, fcs, population, negate = F)
}
\arguments{
\item{densdat}{The full DensityData object with density and densities and set/predicted thresholds.}

\item{exprs}{A n*d expression matrix of the parent population where n is the number of cells and d is the number of channels. The channel indices correspond to the channelIdx values in densdat.}

\item{fcs}{The name of the FCS file to consider. This is needed to filter the relevant entries in densdat.}

\item{population}{The name of the population to gate. This is needed to filter the relevant entries in densdat.}

\item{negate}{Whether the gated population should be negated/inverted, i.e. not the cells within the gate but out of it are considered.}
}
\value{
A vector of length n, indicating cell membership to the population of interest.
}
\description{
This method gates each cell and indicates whether it belongs to the given target population or not. For that, the method needs the original expression matrix.
}
\examples{
f <- unique(flSampleDensdat@data$fcs)[[1]]
flGetGateAssignments(flSampleDensdat, flSampleBcellEvaluationData[[f]]$parentExprs, f, 'bcell')


}
