% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{flIdentifyOutliersF1}
\alias{flIdentifyOutliersF1}
\title{Identify samples outliers by the means of F1-scores.}
\usage{
flIdentifyOutliersF1(f1s)
}
\arguments{
\item{f1s}{The list of F1-scores, one entry per sample.}
}
\value{
A vector of indices with low F1-scores.
}
\description{
This method returns indices of samples with low F1-scores, using an outlier heuristic that considers all samples as outliers, when they are below the 25-percent-quantile - 1.5 * inter-quartile-range.
}
\examples{
f1s <- runif(99, min = 0.42, max = 0.84)
f1s[100] <- 0
flIdentifyOutliersF1(f1s)

}
