% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flInit}
\alias{flInit}
\title{Initializes a DensityData object with default values}
\usage{
flInit(obj)
}
\arguments{
\item{obj}{An uninitialized DensityData object.}
}
\value{
The initialized DensityData Object.
}
\description{
Initializes a DensityData object with default values
}
\examples{
str(flInit(new('DensityData'))) # summarizes an empty DensityData object that has just been initialized

}
