% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flNormalizePopulationName}
\alias{flNormalizePopulationName}
\title{Normalizes population names, such that they can be used, for example in file names.}
\usage{
flNormalizePopulationName(population)
}
\arguments{
\item{population}{The (unnormalized) name of a population.}
}
\value{
The normalized population name.
}
\description{
In particular, '+' is replaced by 'p' (positive), '-' is replaced by 'n' (negative), and all special characters are removed.
}
\examples{
popName <- 'IGD+ CD27-'
popNameNormalized <- flNormalizePopulationName(popName)      
print(popNameNormalized)

}
