% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flNumFeatures}
\alias{flNumFeatures}
\title{Getter for the "numFeatures" slot}
\usage{
flNumFeatures(obj)
}
\arguments{
\item{obj}{A DensityData object.}
}
\value{
The number of features used by this DensityData
}
\description{
Getter for the "numFeatures" slot
}
\examples{
flNumFeatures(flSampleDensdat)
      
}
