% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flPlotDensThresh}
\alias{flPlotDensThresh}
\title{Plots a given density and true/predicted thresholds, if supplied.}
\usage{
flPlotDensThresh(dens, thresh = NULL, predicted = NULL, xlab = "Channel",
  ylab = "Density")
}
\arguments{
\item{dens}{A R density object such as returned from the stats::density function.}

\item{thresh}{The true threshold in the given density. (optional)}

\item{predicted}{The predicted threshold in the given density.}

\item{xlab}{The label of the x-axis (defaults to "Channel")}

\item{ylab}{The label of the y-axis (defaults to "Density")}
}
\value{
Nothing
}
\description{
Plots a given density and true/predicted thresholds, if supplied. True thresholds are drawn as red lines. Predicted thresholds are drawn as blue lines.
If no thresholds are specified, they are not drawn.
}
\examples{
dd <- flFind(flSampleDensdat, 'population == "myeloid" & channelIdx == 1')
protoIdx <- flSelectPrototypes(dd, 1)
ddp <- flPredictThresholds(dd, protoIdx)

par(mfrow = c(3,1))
for(i in 1:3) flPlotDensThresh(flGetDensity(flAt(ddp, i)), flGetGate(flAt(dd, i)), flGetGate(flAt(ddp, i)))

}
