% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flPredictThresholds}
\alias{flPredictThresholds}
\title{Predicts thresholds for all densities in a DensityData object, using a set of prototypes.}
\usage{
flPredictThresholds(densdat, protoIdx)
}
\arguments{
\item{densdat}{A DensityData object containing prototypes and non-prototypes. Prototypes must have thresholds set. All densities in densdat have to be from the same channel.}

\item{protoIdx}{A set of indices that specify prototype rows in densdat.}
}
\value{
A copy of densdat where non-prototype densities have predicted thresholds set.
}
\description{
This method uses the prototypes specified in protoIdx to predict thresholds of all other (non-prototype indices) in densdat.
For each density, it determines the nearest prototype density and calls flAlignThreshold to transfer lower and upper thresholds.
Densities in densdat that are prototypes (specified by protoIdx) must have thresholds set.
Existing non-prototype thresholds in densdat are replaced by predicted thresholds.
}
\examples{
dd <- flFind(flSampleDensdat, population = 'notplasma', channelIdx = 2)
protoIdx <- flSelectPrototypes(dd, 1)
ddp <- flPredictThresholds(dd, protoIdx)

par(mfrow = c(3,1))
for(i in 1:3) flPlotDensThresh(flGetDensity(flAt(ddp, i)), flGetGate(flAt(dd, i)), flGetGate(flAt(ddp, i)))

}
