% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flSampleBcellEvaluationData}
\alias{flSampleBcellEvaluationData}
\title{Example evaluation data matching flSampleDensdat.}
\format{An object of class \code{list} of length 3.}
\usage{
flSampleBcellEvaluationData
}
\description{
Example evaluation data matching flSampleDensdat.
}
\keyword{datasets}
