% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flSampleDensdat}
\alias{flSampleDensdat}
\title{Example DensityData object, containing three samples, 16 populations, and two channels each.}
\format{An object of class \code{DensityData} of length 1.}
\usage{
flSampleDensdat
}
\description{
Example DensityData object, containing three samples, 16 populations, and two channels each.
}
\keyword{datasets}
