% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flSampleFlowFrame}
\alias{flSampleFlowFrame}
\title{Example flowFrame object, used in vignette for constructing DensityData objects.}
\format{An object of class \code{flowFrame} with 135086 rows and 2 columns.}
\usage{
flSampleFlowFrame
}
\description{
Example flowFrame object, used in vignette for constructing DensityData objects.
}
\keyword{datasets}
