% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flSelectPrototypes}
\alias{flSelectPrototypes}
\title{Given a DensityData object, selects a number of prototype densities from it.}
\usage{
flSelectPrototypes(densdat, k)
}
\arguments{
\item{densdat}{The DensityData object. It should contain densities from the same channel only.}

\item{k}{The number of prototypes.}
}
\value{
Vector of indices in the densdat data that were determined to be prototypes.
}
\description{
Using PAM clustering on L1 density distances, this method returns a list of prototype indices which are rows in the given DensityData object.
}
\examples{
dd <- flFind(flSampleDensdat, population = 'notplasma', channelIdx = 2)
protoIdx <- flSelectPrototypes(dd, 1)
print(protoIdx)

}
