% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flSize}
\alias{flSize}
\title{Returns the size of a given DensityData object.}
\usage{
flSize(obj)
}
\arguments{
\item{obj}{The DensityData object.}
}
\value{
The size of obj, i.e. the number of rows of the internal "data" data frame.
}
\description{
Returns the size of a given DensityData object.
}
\examples{
print(flSize(flSampleDensdat)) # print size of flSampleDensdat

}
