% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{gate}
\alias{gate}
\title{Gates a given population using provided thresholds}
\usage{
gate(sample, population, predictedThreshA, predictedThreshB, negate)
}
\arguments{
\item{sample}{The name of the sample for loading the parent population and true gate information.}

\item{population}{The name of the child population.}

\item{predictedThreshA}{A two-element vector containing predicted lower and upper thresholds for density A (or NaN if not exists)}

\item{predictedThreshB}{A two-element vector containing predicted lower and upper thresholds for density B (or NaN if not exists)}
}
\value{
A logical vector with the number cells entries, defining whether a cell is in the gated population or not
}
\description{
Gates a given population using provided thresholds
}

