% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getMeanProportion}
\alias{getMeanProportion}
\title{Reads proportions for a given population and return the mean proportion.}
\usage{
getMeanProportion(tr, population)
}
\arguments{
\item{tr}{The LearningSet object to calculate the statistic for.}

\item{population}{The population name.}
}
\value{
The mean proportion.
}
\description{
Reads proportions for a given population and return the mean proportion.
}

