% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_data.R
\docType{class}
\name{LearningSet-class}
\alias{LearningSet}
\alias{LearningSet-class}
\title{The LearningSet class contains similar entries as TrainingGate,}
\description{
However the latter might contain densities of multiple granularity and evaluation
information such as true gate assignments. The class LearningSet is a more
condensed version containing less information, making it easier to load it as
an intermediate object during evaluation.
}
\section{Slots}{

\describe{
\item{\code{densXchanA}}{Matrix of density x values for channel A for each sample / row}

\item{\code{densXchanB}}{Matrix of density x values for channel B for each sample / row}

\item{\code{densYchanA}}{Matrix of density estimations for channel A for each sample / row}

\item{\code{densYchanB}}{Matrix of density estimations for channel B for each sample / row}

\item{\code{threshA}}{Two-element vector containing lower and upper threshold or NaN for channel A}

\item{\code{threshB}}{Two-element vector containing lower and upper threshold or NaN for channel B}

\item{\code{samples}}{Names of samples}
}}

