% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalizePopulationName}
\alias{normalizePopulationName}
\title{Normalizes a population name.}
\usage{
normalizePopulationName(population)
}
\arguments{
\item{population}{The unnormalized population name string.}
}
\value{
The normalized population string.
}
\description{
First, all '+' and '-' signs are converted to 'p' and 'n', respectively.
Second, All non-characters and non-numbers are removed.
Third, everything is converted to lower case.
}

