% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_data.R
\name{permTrainFiles}
\alias{permTrainFiles}
\title{Permutes a LearningSet object (as loaded by readTrainFiles).}
\usage{
permTrainFiles(tr, perm = NaN, seed = NaN, subsampleRatio = NaN)
}
\arguments{
\item{tr}{The LearningSet object.}

\item{perm}{The permutation to use (use the sample function).}

\item{seed}{Random seed to use (ignored when perm is given).}

\item{subsampleRatio}{Only return this proportion of the given data, i.e. for bootstrapping. Ignored if perm is given}
}
\value{
A LearningSet object that was permuted by the given parameters.
}
\description{
Permutes a LearningSet object (as loaded by readTrainFiles).
}

