% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDensThresh}
\alias{plotDensThresh}
\title{Plots a density and true/predicted thresholds.}
\usage{
plotDensThresh(densX, densY, thresh = NULL, predicted = NULL)
}
\arguments{
\item{densX}{Density x values.}

\item{densY}{Density y values.}

\item{thresh}{True threshold.}

\item{predicted}{Predicted threshold.}
}
\description{
Plots a density and true/predicted thresholds.
}

