% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotGate}
\alias{plotGate}
\title{Plots a gates as given by a TrainingGate object.}
\usage{
plotGate(trainGate, numFeatures = 512)
}
\arguments{
\item{trainGate}{The TrainingGate object to plot.}

\item{numFeatures}{Number of features to use.}
}
\description{
Plots a gates as given by a TrainingGate object.
}

