% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictThresholds}
\alias{predictThresholds}
\title{Predicts thresholds for one population.}
\usage{
predictThresholds(tr, selectedPrototypes)
}
\arguments{
\item{tr}{The LearningSet object.}

\item{selectedPrototypes}{Prototypes as selected by the selectPrototypes() or selectFixedPrototypes() functions.}
}
\value{
A list with the following keys:
\itemize{
 \item{"threshA": A n-by-2 matrix with predicted thresholds for channel A. NaN if threshold does not exist.}
 \item{"threshB": A n-by-2 matrix with predicted thresholds for channel B. NaN if threshold does not exist.}
}
}
\description{
Predicts thresholds for one population.
}

