% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_data.R
\name{readTrainFiles}
\alias{readTrainFiles}
\title{Loads training data for one population.}
\usage{
readTrainFiles(folder, population, numFeatures = 512)
}
\arguments{
\item{folder}{Folder containing the data.}

\item{population}{The population to load.}

\item{numFeatures}{Number of features to use.}
}
\value{
A LearningSet object that contains the requested data.
}
\description{
Loads training data for one population.
}

