% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rotateData}
\alias{rotateData}
\title{Rotates FCS data by theta.}
\usage{
rotateData(data, chans = NULL, theta = NULL)
}
\arguments{
\item{data}{The flowFrame to rotate or the ecpression matrix.}

\item{chans}{Only rotate the given channels.}

\item{theta}{Amount of rotation to apply.}
}
\value{
A rotated version of the input.
}
\description{
Rotates FCS data by theta.
}

