% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{selectFixedPrototypes}
\alias{selectFixedPrototypes}
\title{Selects prototypes for a given set of samples.}
\usage{
selectFixedPrototypes(tr, dA, dB, trainIdxA, trainIdxB)
}
\arguments{
\item{tr}{The LearningSet object.}

\item{dA}{The DTW distances for channel A.}

\item{dB}{The DTW distances for channel B.
@param trainIdxA Prototype indices for channel A.
@param trainIdxB Prototype indices for channel B.}
}
\value{
A list with the following keys:
\itemize{
 \item{"testIdx": Test indices not containing the selected prototype indices.}
 \item{"prototypesA": Prototype indices for channel A.}
 \item{"prototypesB": Prototype indices for channel B.}
}
}
\description{
For now, prototypes are selected randomly.
}

