% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{selectPrototypes}
\alias{selectPrototypes}
\title{Selects prototypes for a given set of samples.}
\usage{
selectPrototypes(tr, dA, dB, numTrain, seed = NaN)
}
\arguments{
\item{tr}{The LearningSet object.}

\item{dA}{The DTW distances for channel A.}

\item{dB}{The DTW distances for channel B.}

\item{numTrain}{The number of prototypes to use (i.e. 2 for 1 per channel)}

\item{seed}{A random seed.}
}
\value{
A list with the following keys:
\itemize{
 \item{"testIdx": Test indices not containing the selected prototype indices.}
 \item{"prototypesA": Prototype indices for channel A.}
 \item{"prototypesB": Prototype indices for channel B.}
}
}
\description{
For now, prototypes are selected randomly.
}

