% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_data.R
\docType{class}
\name{TrainingGate-class}
\alias{TrainingGate}
\alias{TrainingGate-class}
\title{The TrainingGate class is an intermediate representation to which training data is converted.}
\description{
Here, training data might come in several forms (e.g.
manual gates, flowDensity gates) and special conversion scripts are required
to generate objects of this class. These objects are later used for training
and testing.
}
\section{Slots}{

\describe{
\item{\code{parentName}}{Name of the parent population}

\item{\code{channelA}}{Index of first channel in an expression matrix}

\item{\code{channelB}}{Index of second channel in an expression matrix}

\item{\code{thresholdALow}}{Lower threshold of channel A (NaN if not existent)}

\item{\code{thresholdAHigh}}{Upper threshold of channel A (NaN if not existent)}

\item{\code{thresholdBLow}}{Lower threshold of channel B (NaN if not existent)}

\item{\code{thresholdBHigh}}{Upper threshold of channel B (NaN if not existent)}

\item{\code{negate}}{TRUE if the gated population is the inverse of the gate set by the thresholds}

\item{\code{gateAssignments}}{Logical vector of length = numer of cells in the parent population, indicating whether  |a gated cell is inside or outside the target population}

\item{\code{densitiesA}}{List of density objects for channel A as generated by flowLearn's estimateDensity function, for different numbers of features.}

\item{\code{densitiesB}}{List of density objects for channel B as generated by flowLearn's estimateDensity  function, for different numbers of features.}
}}

