% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{weighDtwDistances}
\alias{weighDtwDistances}
\title{Given a DTW distance matrix, weighs it according to Weighted DTW.}
\usage{
weighDtwDistances(m, g = 0.025)
}
\arguments{
\item{m}{Distance matrix, i.e. as generated by dtwDistanceMatrix() or derivativeDtwDistanceMatrix().}

\item{g}{Steepness of the underlying sigmoid function.}
}
\value{
A matrix of the same size of the input matrix, containing weighted elements.
}
\description{
Given a DTW distance matrix, weighs it according to Weighted DTW.
}

