% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowLearn.R
\name{flFind}
\alias{flFind}
\title{Returns a new DensityData object with a subset of entries of the supplied obj.}
\usage{
flFind(obj, mysubset)
}
\arguments{
\item{obj}{The DensityData object to be filtered.}

\item{mysubset}{A logical expression, combining columns of the DensityData object.}
}
\value{
A new DensityData object with a subset of entries, defined by mysubset.
}
\description{
Returns a new DensityData object with a subset of entries of the supplied obj.
}
\examples{
# Get all densities for Population "hfa" and the first channel
dd <- flFind(flSampleDensdat, 'population == "hfa" & channelIdx == 1')
      
}
