% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flSampleBcellEvaluationData}
\alias{flSampleBcellEvaluationData}
\title{Evaluation data to gate B-cells, used in flowLearnVignette.}
\format{An object of class \code{list} of length 3.}
\usage{
flSampleBcellEvaluationData
}
\description{
Evaluation data to gate B-cells, used in flowLearnVignette.
}
\keyword{datasets}
