% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flSampleFlowFrame}
\alias{flSampleFlowFrame}
\title{Example data to show how to construct DensityData objects from flowFrames, used in flowLearnVignette.}
\format{An object of class \code{flowFrame} with 296875 rows and 2 columns.}
\usage{
flSampleFlowFrame
}
\description{
Example data to show how to construct DensityData objects from flowFrames, used in flowLearnVignette.
}
\keyword{datasets}
